SOW-24: split context provider file (resolve lint warning)
