+++ src/context/use-form-context.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { FormContext } from "./FormContext";
+
+export const useFormContext = () => useContext(FormContext);
