+++ src/routes/Page.tsx
@@ -2,7 +2,7 @@ import { Outlet } from "react-router";
 
 import { Header, Footer } from "@lib/gds-components";
 import { PageRoute } from "./routes";
-import { FormProvider } from "../pages/FormContext";
+import { FormProvider } from "../context/FormContext";
 
 export const Page = () => (
   <>
